+++ packages/ng/src/app/user/index.ts
@@ -2,4 +2,5 @@ export { IUser } from './user.model';
 export * from './tile';
 export * from './display';
 export * from './picture';
+export * from './select';
 export { LuUserModule } from './user.module';